Cast every operand in the WPA MAC packs, not only the one that overflows
From the review of #4895. b0fd60373 put the (u32) on the last line of each
group of four, which is the only shift that reaches the sign bit, and that line
then lost the two spaces that hold the shift column against the longer name in
the group below it. The column broke in all twelve groups across m16800,
m16801, m22000 and m22001.

All four operands take the cast now, the padding stays, and the column is where
it was:

    wpa_pmkid->pmkid_data[2] = ((u32) wpa_pmkid->orig_mac_ap[0]  <<  0)
                             | ((u32) wpa_pmkid->orig_mac_ap[1]  <<  8)
                             | ((u32) wpa_pmkid->orig_mac_ap[2]  << 16)
                             | ((u32) wpa_pmkid->orig_mac_ap[3]  << 24);

The shift UndefinedBehaviorSanitizer reported is the one at 24, and it keeps
the cast b0fd60373 gave it, so what the sanitizer sees is unchanged from that
commit; its reproduction still stands. The three other operands shift a byte by
0, 8 and 16, which is defined for any value a MAC byte can hold, so this is the
shape leaving the file rather than a second finding.

All four still crack their own example hash, the two deprecated ones with
--deprecated-check-disable.

+++ src/modules/module_16801.c
@@ -456,17 +456,17 @@ int module_hash_decode (MAYBE_UNUSED const hashconfig_t *hashconfig, MAYBE_UNUSE
 
   wpa_pmkid->pmkid_data[0] = 0x204b4d50; // "PMK "
   wpa_pmkid->pmkid_data[1] = 0x656d614e; // "Name"
-  wpa_pmkid->pmkid_data[2] = (wpa_pmkid->orig_mac_ap[0]  <<  0)
-                           | (wpa_pmkid->orig_mac_ap[1]  <<  8)
-                           | (wpa_pmkid->orig_mac_ap[2]  << 16)
-                           | ((u32) wpa_pmkid->orig_mac_ap[3] << 24);
-  wpa_pmkid->pmkid_data[3] = (wpa_pmkid->orig_mac_ap[4]  <<  0)
-                           | (wpa_pmkid->orig_mac_ap[5]  <<  8)
-                           | (wpa_pmkid->orig_mac_sta[0] << 16)
+  wpa_pmkid->pmkid_data[2] = ((u32) wpa_pmkid->orig_mac_ap[0]  <<  0)
+                           | ((u32) wpa_pmkid->orig_mac_ap[1]  <<  8)
+                           | ((u32) wpa_pmkid->orig_mac_ap[2]  << 16)
+                           | ((u32) wpa_pmkid->orig_mac_ap[3]  << 24);
+  wpa_pmkid->pmkid_data[3] = ((u32) wpa_pmkid->orig_mac_ap[4]  <<  0)
+                           | ((u32) wpa_pmkid->orig_mac_ap[5]  <<  8)
+                           | ((u32) wpa_pmkid->orig_mac_sta[0] << 16)
                            | ((u32) wpa_pmkid->orig_mac_sta[1] << 24);
-  wpa_pmkid->pmkid_data[4] = (wpa_pmkid->orig_mac_sta[2] <<  0)
-                           | (wpa_pmkid->orig_mac_sta[3] <<  8)
-                           | (wpa_pmkid->orig_mac_sta[4] << 16)
+  wpa_pmkid->pmkid_data[4] = ((u32) wpa_pmkid->orig_mac_sta[2] <<  0)
+                           | ((u32) wpa_pmkid->orig_mac_sta[3] <<  8)
+                           | ((u32) wpa_pmkid->orig_mac_sta[4] << 16)
                            | ((u32) wpa_pmkid->orig_mac_sta[5] << 24);
 
   // salt
